Move and start cleaning content property validation
